+++ Magic/src/main/java/com/elmakers/mine/bukkit/magic/ManaController.java
@@ -3,9 +3,9 @@ package com.elmakers.mine.bukkit.magic;
 import org.bukkit.entity.Player;
 
 public interface ManaController {
-    int getMaxMana(Player player);
-    int getManaRegen(Player player);
-    float getMana(Player player);
-    void removeMana(Player player, float amount);
-    void setMana(Player player, float amount);
+    double getMaxMana(Player player);
+    double getManaRegen(Player player);
+    double getMana(Player player);
+    void removeMana(Player player, double amount);
+    void setMana(Player player, double amount);
 }
